update oc.sh to set duration as a parameter
calculate difference between job timestamp and the current date, to check if thejobs has to be deleted

+++ oc.sh
@@ -1,14 +1,3 @@
-#!/bin/sh
-
-export KUBECONFIG=/tmp/.kube
-
-set -e
-
-DEFAULT_NAMESPACE=$(eval cat /var/run/secrets/kubernetes.io/serviceaccount/namespace)
-
-oc login https://$KUBERNETES_PORT_443_TCP_ADDR:$KUBERNETES_SERVICE_PORT_HTTPS \
-  --token `cat /var/run/secrets/kubernetes.io/serviceaccount/token` \
-  --certificate-authority=/var/run/secrets/kubernetes.io/serviceaccount/ca.crt
 oc get jobs -n ${DEFAULT_NAMESPACE} > /tmp/jobs
 tail -n +2 /tmp/jobs > /tmp/jobs-without-header
 
@@ -16,9 +5,21 @@ while read JOB_WITH_STATS
 do
     JOB=$(eval echo "${JOB_WITH_STATS}" | awk '{print $1}')
     SUCCESSFUL_RUN=$(eval echo "${JOB_WITH_STATS}" | awk '{print $3}')
+
     if [ ${SUCCESSFUL_RUN} == "1" ]; then
-      echo "Successfully ended job \"${JOB}\", delete it"
-      oc delete job ${JOB} -n ${DEFAULT_NAMESPACE}
+      START=`oc get job -n abcd1234 -o yaml | grep -i "startTime:" | awk '{print $2}'`
+      START_TIMESTAMP=`date -d "$START" +%s`
+      DATE=`date +%s`
+      echo "start:" $START_TIMESTAMP
+      echo "date:" $DATE
+      DURATION=`expr $DATE - $START_TIMESTAMP`
+        if [[ $DURATION > $1 ]] ;
+          then
+            echo "Successfully ended job \"${JOB}\", delete it"
+            oc delete job ${JOB} -n ${DEFAULT_NAMESPACE}
+          else
+            echo "not to delete"
+          fi
     else
       echo "\"${JOB}\" not successfully ended"
     fi
